In The Wild link tidy up - Art/Photography section
Removed and corrected broken links in the Art/Photography section of OpenSeadragon - In the Wild. This is to improve the experience of the visitors to this page. No effort made at tidying up subsequent sections, however nobody is going to wade through the page if the first link is broken so it is helpful to get the first section tidied.